Int perms

+++ src/api/roles/api.js
@@ -1,6 +1,7 @@
 import Boom from '@hapi/boom';
 import { say } from '../../say';
 import role from './roles';
+import permissions from '../../permissions';
 import ueEvents from '../../events/ueEvents';
 
 const RESOURCE = 'Role';
@@ -19,6 +20,7 @@ const api = {
 	async getAllRoles (req, res, next) {
 		try {
 			if(!req.params.group) throw Boom.preconditionRequired('Must provide authGroup');
+			if(req.permissions.enforceOwn === true) throw Boom.forbidden();
 			const result = await role.getAllRoles(req.authGroup.id, req.query);
 			return res.respond(say.ok(result, RESOURCE));
 		} catch (error) {
@@ -30,6 +32,7 @@ const api = {
 			if(!req.params.group) throw Boom.preconditionRequired('Must provide authGroup');
 			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
 			if (!req.organization) throw Boom.forbidden('Custom roles must be associated to one organization');
+			await permissions.enforceOwnOrg(req.permissions, req.organization.id);
 			const result = await role.getOrganizationRoles(req.authGroup.id, req.product.id, req.organization.id, req.query);
 			return res.respond(say.ok(result, RESOURCE));
 		} catch (error) {
@@ -42,6 +45,7 @@ const api = {
 			if (req.authGroup.active === false) throw Boom.forbidden('You can not add roles in an inactive group');
 			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
 			if (!req.organization) throw Boom.forbidden('Custom roles must be associated to one organization');
+			await permissions.enforceOwnOrg(req.permissions, req.organization.id);
 			if (req.user && req.user.sub) {
 				req.body.createdBy = req.user.sub;
 				req.body.modifiedBy = req.user.sub;
@@ -60,8 +64,9 @@ const api = {
 	// role scoped to authgroup
 	async writeRole(req, res, next) {
 		try {
-			if (req.authGroup.active === false) throw Boom.forbidden('You can not add roles in an inactive group');
-			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
+			if(req.authGroup.active === false) throw Boom.forbidden('You can not add roles in an inactive group');
+			if(!req.product) throw Boom.forbidden('Roles must be associated to one product');
+			if(req.permissions.enforceOwn === true) throw Boom.forbidden();
 			if (req.user && req.user.sub) {
 				req.body.createdBy = req.user.sub;
 				req.body.modifiedBy = req.user.sub;
@@ -79,7 +84,8 @@ const api = {
 	async getRoles(req, res, next) {
 		try {
 			if(!req.params.group) throw Boom.preconditionRequired('Must provide authGroup');
-			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
+			if(!req.product) throw Boom.forbidden('Roles must be associated to one product');
+			await permissions.enforceOwnProduct(req.permissions, req.product.id);
 			const result = await role.getRoles(req.authGroup.id, req.product.id, req.query);
 			return res.respond(say.ok(result, RESOURCE));
 		} catch (error) {
@@ -91,7 +97,7 @@ const api = {
 			if(!req.params.group) throw Boom.preconditionRequired('Must provide authGroup');
 			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
 			if(!req.params.id) throw Boom.preconditionRequired('Must provide id');
-			//todo access - if role is custom, user must have access to the associated organization
+			await permissions.enforceOwnProduct(req.permissions, req.product.id);
 			const result = await role.getRole(req.authGroup.id, req.product.id, req.params.id);
 			if (!result) throw Boom.notFound(`id requested was ${req.params.id}`);
 			return res.respond(say.ok(result, RESOURCE));
@@ -104,8 +110,10 @@ const api = {
 			if(!req.params.group) throw Boom.preconditionRequired('Must provide authGroup');
 			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
 			if(!req.params.id) throw Boom.preconditionRequired('Must provide id');
-			// todo access - if the role is custom, only owner and admin from that org can update
-			const result = await role.patchRole(req.authGroup, req.params.id, req.product.id, req.body, req.user.sub || req.user.id || 'SYSTEM');
+			if(req.permissions.enforceOwn === true) throw Boom.forbidden();
+			const thisRole = await role.getRole(req.authGroup.id, req.product.id, req.params.id);
+			if(thisRole.core === true) await permissions.enforceRoot(req.permissions);
+			const result = await role.patchRole(req.authGroup, thisRole, req.params.id, req.product.id, req.body, req.user.sub || req.user.id || 'SYSTEM');
 			return res.respond(say.ok(result, RESOURCE));
 		} catch (error) {
 			ueEvents.emit(req.authGroup.id, 'ue.role.error', error);
@@ -117,7 +125,9 @@ const api = {
 			if(!req.params.group) throw Boom.preconditionRequired('Must provide authGroup');
 			if (!req.product) throw Boom.forbidden('Roles must be associated to one product');
 			if(!req.params.id) throw Boom.preconditionRequired('Must provide id');
-			// todo access - if role is custom, user must be from the org
+			if(req.permissions.enforceOwn === true) throw Boom.forbidden();
+			const thisRole = await role.getRole(req.authGroup.id, req.product.id, req.params.id);
+			if(thisRole.core === true) await permissions.enforceRoot(req.permissions);
 			const result = await role.deleteRole(req.authGroup.id, req.product.id, req.params.id);
 			if (!result) throw Boom.notFound(`id requested was ${req.params.id}`);
 			return res.respond(say.ok(result, RESOURCE));
